move potentially more generic classes up to .Common from .Http, ditch IDiscoveryLoadBalancer
