+++ src/Steeltoe.Common.Http/Discovery/DiscoveryHttpClientHandler.cs
@@ -45,7 +45,7 @@ namespace Steeltoe.Common.Discovery
             var current = request.RequestUri;
             try
             {
-                request.RequestUri = _discoveryBase.LookupService(current);
+                request.RequestUri = await _discoveryBase.LookupServiceAsync(current);
                 return await base.SendAsync(request, cancellationToken);
             }
             catch (Exception e)
