+++ src/Steeltoe.Common/Discovery/ConfigurationServiceInstanceProvider.cs
