+++ src/Steeltoe.Common/Discovery/ConfigurationServiceInstanceProviderServiceCollectionExtensions.cs
