+++ src/Steeltoe.Common/LoadBalancer/RandomLoadBalancer.cs
@@ -14,11 +14,10 @@
 
 using Microsoft.Extensions.Logging;
 using Steeltoe.Common.Discovery;
-using Steeltoe.Common.LoadBalancer;
 using System;
 using System.Threading.Tasks;
 
-namespace Steeltoe.Common.Http.LoadBalancer
+namespace Steeltoe.Common.LoadBalancer
 {
     public class RandomLoadBalancer : ILoadBalancer
     {
