+++ src/Steeltoe.Common/LoadBalancer/RoundRobinDistributedLoadBalancer.cs
@@ -15,12 +15,11 @@
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Logging;
 using Steeltoe.Common.Discovery;
-using Steeltoe.Common.LoadBalancer;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Steeltoe.Common.Http.LoadBalancer
+namespace Steeltoe.Common.LoadBalancer
 {
     public class RoundRobinDistributedLoadBalancer : ILoadBalancer
     {
